+++ cpp/dolfinx/fem/petsc.cpp
@@ -9,7 +9,7 @@
 #include "assembler.h"
 #include "sparsitybuild.h"
 #include <dolfinx/common/IndexMap.h>
-#include <dolfinx/la/PETScMatrix.h>
+#include <dolfinx/la/petsc.h>
 #include <dolfinx/la/SparsityPattern.h>
 #include <xtl/xspan.hpp>
