+++ cpp/dolfinx/la/CMakeLists.txt
@@ -1,6 +1,6 @@
 set(HEADERS_la
   ${CMAKE_CURRENT_SOURCE_DIR}/dolfin_la.h
-  ${CMAKE_CURRENT_SOURCE_DIR}/PETScMatrix.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/petsc.h
   ${CMAKE_CURRENT_SOURCE_DIR}/SLEPcEigenSolver.h
   ${CMAKE_CURRENT_SOURCE_DIR}/SparsityPattern.h
   ${CMAKE_CURRENT_SOURCE_DIR}/utils.h
@@ -8,7 +8,7 @@ set(HEADERS_la
   PARENT_SCOPE)
 
 target_sources(dolfinx PRIVATE
-  ${CMAKE_CURRENT_SOURCE_DIR}/PETScMatrix.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/petsc.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/SLEPcEigenSolver.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/SparsityPattern.cpp
 )
