+++ cpp/dolfinx/la/petsc.cpp
@@ -5,7 +5,7 @@
 //
 // SPDX-License-Identifier:    LGPL-3.0-or-later
 
-#include "PETScMatrix.h"
+#include "petsc.h"
 #include "Vector.h"
 #include "utils.h"
 #include <dolfinx/common/IndexMap.h>
